Make it so that N216_ens_analysis can produce all_cases figures.
Quite a lot of work to get this working. One of the files (u-di728 04
Z500) has an error that means it's got no metadata - it took a while to
track this down. I tried fixing it to no avail, so am ignoring April for
now. Combining the datasets involved resetting the time dim to be the
same for each ds, so that I could concat each across a new dim, case.

Now producing similar plots to before but averaged over all cases.